Move options to commands directory
