+++ src/commands/help.ts
