+++ src/commands/version.ts
